+++ CHANGES.md
@@ -4,6 +4,7 @@
 
 * allow `API_DISABLE_{COG/STAC/MOSAIC}` environment variables to control default endpoints in titiler main app (ref: https://github.com/developmentseed/titiler/issues/156)
 * add `overwriting=False/True` on MosaicJSON creation (ref: https://github.com/developmentseed/titiler/issues/164)
+* add `gdal_config` option to Tiler factories to replace custom `APIRoute` class (ref: https://github.com/developmentseed/titiler/issues/168)
 
 ## 0.1.0a12 (2020-11-18)